+++ hyperdbg/hprdbgkd/code/debugger/user-level/ThreadHolder.c
@@ -88,7 +88,7 @@ ThreadHolderIsAnyPausedThreadInProcess(PUSERMODE_DEBUGGING_PROCESS_DETAILS Proce
 
         for (size_t i = 0; i < MAX_THREADS_IN_A_PROCESS_HOLDER; i++)
         {
-            if (ThreadHolder->Threads[i].ThreadId != NULL && ThreadHolder->Threads[i].IsPaused)
+            if (ThreadHolder->Threads[i].ThreadId != (UINT32)NULL && ThreadHolder->Threads[i].IsPaused)
             {
                 return TRUE;
             }
@@ -179,7 +179,7 @@ ThreadHolderGetProcessFirstThreadDetailsByProcessId(UINT32 ProcessId)
 
         for (size_t i = 0; i < MAX_THREADS_IN_A_PROCESS_HOLDER; i++)
         {
-            if (ThreadHolder->Threads[i].ThreadId != NULL)
+            if (ThreadHolder->Threads[i].ThreadId != (UINT32)NULL)
             {
                 //
                 // The active thread's structure is found
@@ -420,7 +420,7 @@ ThreadHolderApplyActionToPausedThreads(PUSERMODE_DEBUGGING_PROCESS_DETAILS Proce
 
             for (size_t i = 0; i < MAX_THREADS_IN_A_PROCESS_HOLDER; i++)
             {
-                if (ThreadHolder->Threads[i].ThreadId != NULL &&
+                if (ThreadHolder->Threads[i].ThreadId != (UINT32)NULL &&
                     ThreadHolder->Threads[i].IsPaused)
                 {
                     for (size_t j = 0; j < MAX_USER_ACTIONS_FOR_THREADS; j++)
